# api/openapi.yaml
openapi: 3.0.3
info:
  title: Todo App API
  description: OpenAPI specification for a simple todo list app.
  version: 1.0.0
servers:
  - url: "http://localhost:8080/api/v1"
    description: Development
paths:
  /todos:
    get:
      tags: [ todos ]
      summary: Retrieves the entire list of todo items.
      operationId: listTodoItems
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "schema/todo_item.yaml"
    post:
      tags: [ todos ]
      summary: Creates a new todo item.
      operationId: createTodoItem
      requestBody:
        description: The todo item to create.
        content:
          application/json:
            schema:
              $ref: "schema/todo_item.yaml"
        required: true
      responses:
        201:
          description: The todo item was created.
          content:
            application/json:
              schema:
                $ref: "schema/todo_item.yaml"
        400:
          description: The request to create the todo item is not valid.
  /todos/{id}:
    get:
      tags: [ todos ]
      summary: Retrieves a todo item by its ID.
      operationId: fetchTodoItem
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "schema/todo_item.yaml"
        404:
          description: No todo item exists with the given ID.
    put:
      tags: [ todos ]
      summary: Replaces a todo item.
      operationId: replaceTodoItem
      responses:
        200:
          description: The todo item was replaced.
          content:
            application/json:
              schema:
                $ref: "schema/todo_item.yaml"
        400:
          description: The request to replace the todo item is not valid.
        404:
          description: No todo item exists with the given ID.
    patch:
      tags: [ todos ]
      summary: Updates a todo item.
      operationId: updateTodoItem
      responses:
        200:
          description: The todo item was updated.
          content:
            application/json:
              schema:
                $ref: "schema/todo_item.yaml"
        400:
          description: The request to update the todo item is not valid.
        404:
          description: No todo item exists with the given ID.
    delete:
      tags: [ todos ]
      summary: Deletes a todo item.
      operationId: deleteTodoItem
      responses:
        204:
          description: The todo item was deleted.
        404:
          description: No todo item exists with the given ID.
    parameters:
      - name: id
        in: path
        description: The ID of a todo item.
        required: true
        schema:
          type: integer
          minimum: 1